backend/requirements.txt
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
Click==7.0
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
future==0.17.1
idna==2.8
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pyasn1==0.4.7
pycryptodome==3.3.1
pylint==2.3.1
python-jose==3.0.1
python-jose-cryptodome==1.3.2
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
urllib3==1.25.7
Werkzeug==0.15.2
wrapt==1.11.1
